cups: Use the same name-mangling as Debian 11's cups-browsed
based in debian gtk patch:
https://salsa.debian.org/gnome-team/gtk3/-/commit/9c79d97423bfb78d48290f006f46376437487e6f

From: Simon McVittie <smcv@debian.org>
Date: Sun, 20 Feb 2022 21:23:25 +0000
Subject: cups: Use the same name-mangling as Debian 11's cups-browsed

The name-mangling used in GTK 3.24.25 is appropriate for
cups-browsed >= 1.28.11, which eliminates leading and trailing
underscores in the normalized name, but the cups-browsed version in
Debian 11 is older and does not do that. We need to match the printer
names generated by the installed cups-browsed, to avoid a printer
showing up twice under confusingly similar names.

Signed-off-by: Simon McVittie <smcv@debian.org>
Forwarded: not-needed, specific to Debian 11 software versions